requirements.txt
python-dotenv
langchain-pinecone
cohere
langchain
langchain-community
langchain-core 
langchain-google-genai
fastapi
uvicorn
transformers
accelerate
gTTS
jinja2
python-multipart
